+++ R/tq_get.R
@@ -961,13 +961,15 @@ tq_get_rblpapi <- function(x, get, rblpapi_fun = "bdh", complete_cases, map, ...
 
     ret <- tryCatch({
 
-        # Repurpose from and to as start_date and end_date
+        # Rblpapi argument setup
         if (rblpapi_fun == "bds") {
             args <- list(security = x)
         } else {
             args <- list(securities = x)
         }
         args <- append(args, list(...))
+
+        # Repurpose from and to as start_date and end_date
         if (!is.null(args$from)) {
             args$start.date <- as.Date(args$from)
             args$from <- NULL
@@ -977,6 +979,7 @@ tq_get_rblpapi <- function(x, get, rblpapi_fun = "bdh", complete_cases, map, ...
             args$to <- NULL
         }
 
+        # Call function
         do.call(rblpapi_fun, args) %>%
             tibble::as.tibble()